requirements.txt
streamlit
google-genai
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
pypdf
speechrecognition
pyttsx3